Appbar를 통해 화면 전환 성공, 화면 전환 후 새로고침 하면 404 Error 발생하는 것을 확인
