+++ src/main/resources/application.properties
@@ -1,4 +1,4 @@
 server.port=8090
 
-spring.mvc.view.prefix=/jsp/
-spring.mvc.view.suffix=.jsp
+spring.mvc.view.prefix=/html/
+spring.mvc.view.suffix=.html
