+++ src/main/webapp/css/custom.css
@@ -1,21 +1,8 @@
 @charset "UTF-8";
-.login {
+.center {
 	position:absolute;
-	top:0;
-	left:0;
-	width:100%;
-	height: 100%;
+	top: 50%;
+	left: 50%;
+	transform: translate(-50%, -50%);
 	text-align: center;
-}
-
-.login .content {
-	display: inline-block;
-	vertical-align: middle;
-}
-
-.login .blank {
-	display: inline-block;
-	width: 0;
-	height: 50%;
-	vertical-align: midlle;
 }
