+++ src/main/webapp/html/index.html
@@ -1,0 +1,11 @@
+<!DOCTYPE html>
+<html>
+<head>
+<meta charset="UTF-8">
+<title></title>
+</head>
+<body>
+	<div id="root"></div>
+	<script src="/bundle/bundle.js"></script>
+</body>
+</html>
