--- src/main/webapp/jsp/page.jsp
@@ -1,13 +1,0 @@
- <%@ page language="java" contentType="text/html; charset=UTF-8"
-    pageEncoding="UTF-8"%>
-<!DOCTYPE html>
-<html>
-<head>
-<meta charset="UTF-8">
-<title>${pageName}</title>
-</head>
-<body>
-	<div id="root"></div>
-	<script src="/bundle/${pageName}.bundle.js"></script>
-</body>
-</html>
